+++ test/test_csvhelper.py
@@ -39,6 +39,8 @@ class TestCSVHelperCase(unittest.TestCase):
         self.assertEquals(len(id_name_dict), 7)
         ''' TODO
         why there dict not support use 0 as key value???
+        python3 dict keys not list type
         '''
-        self.assertEquals(id_name_dict.get(id_name_dict.keys()[0]), '中国')
+        self.assertEquals(id_name_dict.get(list(id_name_dict.keys())[0]), '中国')
+        self.assertEquals(id_name_dict.get('0'), '中国')
         self.assertEquals(id_name_dict.get('1'), 'Python')